+++ src/data/list/array/vec/reexports.rs
@@ -4,4 +4,4 @@ use crate::{impl_cdef, mod_path};
 
 mod_path!(alloc +pub _a "../../../../../libs/base_alloc/src/data/list/array/vec/reexports.rs");
 
-impl_cdef![<T> Self::new() => Vec<T>]; // impl ConstDefault
+impl_cdef![ConstDefault: <T> Self::new() => Vec<T>]; // impl ConstDefault