+++ src/data/list/queue/reexports.rs
@@ -3,4 +3,4 @@
 crate::mod_path!(alloc +pub _a "../../../../libs/base_alloc/src/data/list/queue/reexports.rs");
 
 #[cfg(feature = "alloc")]
-crate::impl_cdef![<T> Self::new() => VecDeque<T>]; // impl ConstDefault
+crate::impl_cdef![ConstDefault: <T> Self::new() => VecDeque<T>]; // impl ConstDefault